Add smoothly-modal (#927)
* Add smoothly-modal

- Dialog:
A dialog is a general term for a container that presents information or prompts user interaction, without necessarily restricting the user’s ability to interact with the rest of the page.

- Modal
A modal is a type of dialog that enforces a blocking interaction pattern, requiring the user to engage with it before returning to the main content.

* Add role

* No div around content

---------

Co-authored-by: qijieye <86956109+qijieye@users.noreply.github.com>